Danish Module of the 2009 European Election Study (EES)

The denmarkEES2009 dataset contains data from the Danish module of the 2009 European Election Study (EES). This dataset includes responses from 1,000 Danish participants who rated their propensity to vote for each of eight political parties on a 0–10 point scale. A score of 0 denotes "not at all possible," while a score of 10 denotes "very probable."

Usage

data(denmarkEES2009)

Format

A matrix with 1,000 rows and 8 columns:

V1

The propensity to vote for Party 1.

V2

The propensity to vote for Party 2.

V3

The propensity to vote for Party 3.

V4

The propensity to vote for Party 4.

V5

The propensity to vote for Party 5.

V6

The propensity to vote for Party 6.

V7

The propensity to vote for Party 7.

V8

The propensity to vote for Party 8.

Details

The dataset can be used to demonstrate the smacofRect() function.

There are only 61 missing ratings in the 1000×8 matrix.

The dataset is particularly useful for demonstrating the use of the smacofRect() function in multidimensional scaling analysis.

See also

Examples

data(denmarkEES2009)