Roll Call Data from the French Fourth Republic

The france4 dataset contains roll call data from the French Fourth Republic, as analyzed by Rosenthal and Voeten (2004). This dataset was used to estimate a party-switcher model where a separate ideal point is estimated each time a legislator changes party affiliation. The dataset includes legislator-specific variables and roll call votes. Rosenthal and Voeten (2004) found that the latent ideological space remained stable over the course of the French Fourth Republic, so the roll call data is not segmented by legislative session.

data(france4)

Format

A data frame with the following variables:

NAME

Name of the deputy (legislator).

MID

Deputy ID, which remains constant even if the deputy switches party.

CASEID

Deputy ID, which is unique for each party affiliation if the deputy switches party.

PAR

Party affiliation of the deputy.

PARSEQ

Sequence of party affiliation for deputies who switched parties.

VOTE1

Result of the first roll call vote (and so on for subsequent votes).

...

Other variables representing additional roll call votes.

Source

Data from Rosenthal and Voeten (2004). The original dataset and documentation can be found online at http://www9.georgetown.edu/faculty/ev42/france.htm.

Details

The dataset includes 1,416 separate ideal points, as each party-switching deputy has a separate entry for each party affiliation. The first five columns of the dataset are legislator-specific variables: NAME (deputy name), MID (constant deputy ID), CASEID (unique ID for each party affiliation), PAR (party affiliation), and PARSEQ (sequence number of party affiliations for party-switching deputies). The remaining columns represent the roll call votes.

The dataset and its extensive documentation were originally made available online by Rosenthal and Voeten (2004). For more information and access to the original data, visit http://www9.georgetown.edu/faculty/ev42/france.htm.

Examples

if (FALSE) { # \dontrun{
data(france4)
} # }