1981 Interest Group Rating Scores

Subset of Keith Poole's Interest Group data (1959-1981)

data(interest1981)

Format

A data frame with 100 observations on the following 49 variables.

year

Year of observation (only 1981)

congress

Congress number (only 97)

session

Session number (only 1)

senate

Senate number (only 1)

prid

ICPSR ID Corrected by Poole and Rosenthal

party

Party Code (1=Democrat, 2=Republican)

state

ICPSR State Code

name

Legislator Name

ada

Americans for Democrat Action Rating Score

cope

Committee on Political Action (AFL-CIO) Rating Score

nfu

National Farmer's Union Rating Score

ccus

Chamber of Commerce of the United States Rating Score

aca

Americans for Constitutional Action Rating Score

cc

Conservative Coalition (CQ) Rating Score

ll

Liberty Lobby Rating Score

asc

American Security Council Rating Score

ripon

Ripon Society Rating Score

uaw

United Auto Workers Rating Score

nea

National Education Association Rating Score

lwv

League of Women Voters Rating Score

ntu

National Taxpayers' Union Rating Score

lcv

League of Conservation VotersRating Score

cfa

Consumer Federation of America Rating Score

nfo

National Farmers Organization Rating Score

aft

American Federation of Teachers Rating Score

nfib

National Federation of Independent Busioness Rating Score

cfnfmp

Coalition for a New Foreign and Military Policy Rating Score

nasc

National Alliance of Senior Citizens Rating Score

afbf

American Farm Bureau Federation Rating Score

ccause

Common Cause Rating Score

bfw

Bread for the World Rating Score

umw

United Mine Workers Rating Score

bctd

Building and Construction Trade Department (AFL-CIO) Rating Score

cv

Christian Voice Rating Score

aclu

American Civil Liberties Union Rating Score

csfca

Committee for the Survival of a Free Congress: All Issues Rating Score

cvvf

Christian Voters Victory Fund Rating Score

reagan

Reagan Support Score (CQ) Rating Score

stateid

Numeric State ID

district

Numeric vector of zeros

surname

Legislator Surname and first initial

dw1

First dimension DW Nominate Score

dw2

Second Dimension DW Nominate Score

se1

SE of DW Nominate First Dimension

se2

SE of DW Nominate Second Dimension

logl

logl

choices

choices

errors

errors

gmp

gmp

Source

http://www.voteview.com/dwnl_5.htm

References

Include Reference