Mexican Module of the Comparative Study of Electoral Systems (CSES) 2000 and 2006

The mexicoCSES2006 dataset contains data from the 2000 and 2006 Mexican modules of the Comparative Study of Electoral Systems (CSES). In these surveys, Mexican citizens were asked to place the major political parties on an 11-point left-right scale.

Usage

data(mexicoCSES2006)

Format

A data frame with rows representing respondents and 8 columns representing different political parties:

PAN

Numeric, placement of the PAN (National Action Party) on the left-right scale.

PRD

Numeric, placement of the PRD (Party of the Democratic Revolution) on the left-right scale.

PRI

Numeric, placement of the PRI (Institutional Revolutionary Party) on the left-right scale.

Greens

Numeric, placement of the Green Party on the left-right scale.

PT

Numeric, placement of the PT (Labor Party) on the left-right scale.

Convergencia

Numeric, placement of the Convergencia (Convergence) party on the left-right scale.

Nueva Alianza

Numeric, placement of the Nueva Alianza (New Alliance) party on the left-right scale.

PSD

Numeric, placement of the PSD (Social Democratic Party) on the left-right scale.

Source

Comparative Study of Electoral Systems (CSES), Mexican modules for 2000 and 2006.

Details

The dataset includes responses from the 2000 and 2006 Mexican modules of the Comparative Study of Electoral Systems (CSES), where respondents were asked to place major political parties on an 11-point left-right scale.

Examples

if (FALSE) { # \dontrun{
data(mexicoCSES2006)
} # }