MLSMU6

Metric unfolding using the MSLMU6 algorithm.

mlsmu6(input, ndim=2, cutoff=5, tol=0.0005, maxit=50, id=NULL)

Arguments

input
An data frame or matrix of individuals by stimuli placements.

ndim
Number of latent dimensions to be generated.

cutoff
Minimum number of stimuli that need to be rated to be included in result.

tol
Tolerance to identify convergence on the aggregate sum of squared errors.

maxit
Maximum number of iterations permitted

id
Optional vector of identifiers (e.g., party labels) for the individuals.

Value

A list that will include the following:

stims

A number of stimuli by number of dimensions matrix of scaled stimulus values.

inds

A number of individuals by number of dimensions matrix of scaled individual values.

iter

Iteration history.