Issue scale data from the 1968 American National Election Study regarding Vietnam

Respondent placements of themselves and political figures on a 1-7 scale of opinions about America's involvement in Vietnam. Endpoints of the scale are ‘Immediate Withdrawal’ and ‘Complete Military Victory’. See ‘Details’ for full question wording.

data(nes1968_urbanunrest)

Format

A numeric matrix of respondent placements with the following column names:

vote.choice

Johson

Respondent placement of Lyndon Johnson (President)

Humphrey

Respondent placement of Hubert Humphrey (D Presidential Candidate)

Nixon

Respondent placement of Richard Nixon (R Presidential Candidate)

Wallace

Respondent placement of George Wallace (I Presidential Candidate)

self

Respondent placement of Self

Details

The question asked in the ANES was as follows:

There is much talk about ‘hawks’ and ‘doves’ in connection with Vietnam, and considerable disagreement as to what action the United States should take in Vietnam. Some people think we should do everything necessary to win a complete military victory, no matter what results. Some people think we should withdraw completely from Vietnam right now, no matter what results. And, of course, other people have opinions somewhere between these two extreme positions. Suppose the people who support an immediate withdrawal are at one end of this scale at point number 1. And suppose the people who support a complete military victory are at the other end of the scale at point number 7. At what point on the scale would you place ...?

Source

The Political Behavior Program of the Survey Research Center, Institute for Social Research, University of Michigan. NATIONAL ELECTION STUDIES, 1968: PRE-/POST-ELECTION STUDY [dataset]. Ann Arbor, MI: University of Michigan, Center for Political Studies [producer and distributor], 1999.