Plot Method for mlsmu6 Class Objects

This function creates a 2D plot of the results from the mlsmu6 function. It visualizes the estimated positions of both individuals and stimuli in a lower-dimensional space.

Usage

# S3 method for class 'mlsmu6'
plot(x, ..., selected.stims = NULL, ind.id.size = 3, stim.id.size = 6)

Arguments

x

An object of class mlsmu6, typically the output from the mlsmu6 function.

...

Additional arguments passed to the plot method (currently not used).

selected.stims

A vector of stimulus names to highlight in the plot. If NULL, all stimuli are plotted.

ind.id.size

Numeric value specifying the text size for individual labels in the plot. Default is 3.

stim.id.size

Numeric value specifying the text size for stimulus labels in the plot. Default is 6.

Value

A ggplot object representing the 2D plot of the estimated positions.

Details

The plot shows individuals and stimuli in a 2D space defined by the first two dimensions (Dim1 and Dim2). Individuals are represented with smaller labels, while stimuli are larger, making them more prominent. If selected.stims is specified, only the selected stimuli will be highlighted, while the rest will be omitted from the plot.

Examples

if (FALSE) { # \dontrun{
# Assuming `result` is the output from `mlsmu6` function:
data(mlsmu6_out)
plot.mlsmu6(mlsmu6_out) 
} # }