Roll Call Data for U.S. Congress

The rc_ep object is a roll call dataset compiled by Poole and Rosenthal. This dataset represents roll call votes in the U.S. Congress and has been processed into an object of class rollcall() for analysis.

data(rc_ep)

Format

An object of class "rollcall" containing roll call vote data.

Source

Poole, K. T., & Rosenthal, H. (Various Years). House and Senate Roll Call Data. Retrieved from http://www.voteview.com.

Details

Poole and Rosenthal have compiled House and Senate roll call datasets covering the history of the U.S. Congress. These datasets are maintained at http://www.voteview.com. The rc_ep object represents a specific subset of these data and is formatted as a rollcall object, suitable for various forms of legislative analysis.

Examples

if (FALSE) { # \dontrun{
data(rc_ep)
} # }