Result of Aldmck Analysis for France EES 2009 Data

This dataset contains the result of the aldmck analysis applied to the franceEES2009 dataset.

Usage

data(result.france)

Format

An object of class "aldmck".

Details

The result.france object is generated using the aldmck function with the following parameters: result.france <- aldmck(franceEES2009, respondent = 1, polarity = 2, missing = c(77, 88, 89), verbose = FALSE)