Swedish Parliamentary Candidate Survey (2010)

2010 Survey of Swedish Parliamentary Candidates

data(Sweden2010)

Format

A data frame with 2751 observations on the following 56 variables. Each of the policy statements is coded 1 (strongly disagree), 2 (disagree), 3 (agree), 4 (strongly agree), 8 (don't know)

id

Candidate ID

elected

Whether the candidate was elected (1) or not (0)

party.name

Name of candidate's party

party.code

Candidate's party code

govt.party

Whether the candidate is in a government party (1) or not (0)

left.right.self.fivept

Candidate's left-right self-placement from (1=left to 5=right)

congestion.taxes

Congestion taxes should be introduced in more big cities

highspeed.trains

Sweden should invest in high-speed trains

hunt.wolves

Sweden should allow license hunt of wolves

nuclear.power

The nuclear power should be further expanded

gasoline.taxes

The tax on gasoline should be raised

museum.fees

The entrance fee for state-owned museums should be removed

online.piracy

It should be a criminal offense to download copyright protected material from the internet for personal use

state.TV

The fee for state-owned TV should be replaced with a mandatory tax

No municipality should be allowed to refuse to take in refugees
refugee.cities

No municipality should be allowed to refuse to take in refugees

asylum.seekers

Fewer asylum seekers should be accepted into Sweden

refugee.healthcare

Refugees without documents should be able to receive free health care

teacher.veils

Teachers should not be allowed to wear face-covering veil in Swedish schools

paternal.leave

A bigger share of the parental leave should be reserved for the fathers

affirmative.action.universities

The use of affirmative action in university admission should be abolished

child.raising.allowance

The child-raising allowance should be abolished

property.taxes.wealthy

The property tax for detached houses with a taxable value above 4.5 millions should be raised

wealth.tax

The wealth tax should be reinstated

tax.wealthy

The tax rate for those with high incomes should be raised

tax.pensions

The same tax rate should be applied to pensions and wages

household.services.deduction

The tax deduction for household services [RUT] should be kept

work.income.tax

The tax on work-income should be lowered

sex.purchase

The sex purchase law should be abolished

DUI.penalty

The penalty for driving under intoxication should be increased

criminal.sentences

Violent crimes should lead to tougher sentences

wiretaps

The FRA law [a legislative package that authorizes the state to wiretap all telephone and Internet traffic that crosses Sweden's borders] should be torn up

retirement.age

Those who are 58 years old should be eligible for early retirement

health.insurance.time

The maximum time on health insurance should be limited

dental.insurance

Dental care should be included in health care insurance

competition.public.sector

There should be more competition in the public sector

mandatory.unemployment.insurance

Unemployment insurance should be mandatory

municipal.home.care

Municipal home-care service fees should be fully equalized across municipalities

circumcision

Swedish health care should only offer to carry out circumcision when it has a medical value

private.healthcare.profits

Government financed private health care service providers should be able to distribute profits to their owners

govt.alcohol.monopoly

The government monopoly on selling alcoholic beverages should be removed

employment.protection

It should be made more difficult for personnel agencies to circumvent employment protection legislation

sell.public.corporations

State-owned corporations should not be sold

Aghanistan.withdrawal

Swedish soldiers in Afghanistan should be brought immediately

exporting.arms

Swedish should cease exporting arms immediately

aid.undemocratic.countries

Foreign aid should not go to undemocratic countries

compulsory.military.service

The compulsory military service should be reinstated

leave.EU

Sweden should leave the European Union

transfer.students

It should be possible to transfer students with discipline problems from their school against the will of the student and their parents

local.control.education

Responsibility for primary and secondary education should be moved from municipalities to the national government

student.grades

Students should received grades from garde six and onwards

number.private.schools

The number of private schools should be limited

university.eligibility

All who pass secondary school programs should be eligible for university studies

criminalize.racist.organizations

It should be a criminal offense to be a member of a racist organisation

abolish.monarchy

Monarchy should be abolished

ballot.order

Voter control over the order of candidates on ballots in Swedish elections should be strengthened

referenda.elections

More referenda should be held in Sweden

Source

Need source citation