MDS Solution from mlsmu6

The mlsmu6_out dataset contains the multidimensional scaling (MDS) solution generated by applying the mlsmu6 function to a subset of the interest1981 dataset. The MDS solution is computed using two dimensions and a cutoff value of 5, with data grouped by political party labels.

Format

A matrix or data frame, depending on the structure of the output from mlsmu6, typically containing coordinates in the reduced dimensional space.

Source

Generated using the mlsmu6 function on the interest1981 dataset.

Details

The mlsmu6_out object is generated by running the following command:

mlsmu6_out <- mlsmu6(input = interest1981[, 9:38], ndim = 2, cutoff = 5,
                     id = factor(interest1981$party, labels = c("D", "R")))

The mlsmu6 function applies multidimensional scaling to reduce the dimensionality of the input data, with the number of dimensions set to 2 and a cutoff value of 5. The input data is a subset of the interest1981 dataset, specifically columns 9 to 38, grouped by the political party (party) variable.

See also

Examples

if (FALSE) { # \dontrun{
data(mlsmu6_out)
} # }